extract strings to its own class and booya
