+++ lib/bashly/views/command/parse_args_case.erb
@@ -8,10 +8,10 @@
 <%- condition = "elif" -%>
 <%- end -%>
 else
-  echo -e "<%= Settings.strings["invalid_argument"] % { arg: "$key" } %>"
+  echo -e "<%= strings["invalid_argument"] % { arg: "$key" } %>"
   exit 1
 fi
 <%- else -%>
-echo -e "<%= Settings.strings["invalid_argument"] % { arg: "$key" } %>"
+echo -e "<%= strings["invalid_argument"] % { arg: "$key" } %>"
 exit 1
 <%- end -%>
