+++ lib/bashly/views/command/parse_args_while.erb
@@ -8,7 +8,7 @@ while [[ $# -gt 0 ]]; do
   <%- end -%>
 
   -* )
-    echo -e "<%= Settings.strings["invalid_flag"] % { flag: "$key" } %>"
+    echo -e "<%= strings["invalid_flag"] % { flag: "$key" } %>"
     exit 1
     ;;
