+++ lib/bashly/views/command/required_flags_filter.erb
@@ -4,7 +4,7 @@ argstring="$*"
 <%- end -%>
 <%- required_flags.each do |flag| -%>
 if [[ <%= flag.aliases.map { |a| %Q["$argstring" != *#{a}*] }.join " && " %> ]]; then
-  echo -e "<%= Settings.strings["missing_required_flag"] % { usage: flag.usage_string } %>"
+  echo -e "<%= strings["missing_required_flag"] % { usage: flag.usage_string } %>"
   exit 1
 fi
 <%- end %>
