+++ lib/bashly/views/command/usage.erb
@@ -2,13 +2,13 @@
 <%= full_name.to_underscore %>_usage() {
   echo -e "<%= caption_string %>"
   echo 
-  echo -e "<%= Settings.strings["usage"] %>"
+  echo -e "<%= strings["usage"] %>"
   echo -e "  <%= usage_string %>"
   echo
 <%= render(:usage_commands).indent 2 if commands.any? %>
 
   if [[ -n $long_usage ]]; then
-    echo -e "<%= Settings.strings["options"] %>"
+    echo -e "<%= strings["options"] %>"
 <%= render(:usage_fixed_flags).indent 4 %>
 <%= render(:usage_flags).indent 4 if flags.any? %>
 <%= render(:usage_args).indent 4 if args.any? %>
