+++ lib/bashly/views/command/usage_args.erb
@@ -1,5 +1,5 @@
 # :command.usage_args
-echo -e "<%= Settings.strings["arguments"] %>"
+echo -e "<%= strings["arguments"] %>"
 
 <%- args.each do |arg| -%>
 <%= arg.render(:usage) %>
