+++ lib/bashly/views/command/usage_commands.erb
@@ -1,5 +1,5 @@
 # :command.usage_commands
-echo -e "<%= Settings.strings["commands"] %>"
+echo -e "<%= strings["commands"] %>"
 <%- maxlen = command_names.map(&:size).max -%>
 <%- commands.each do |command| -%>
 echo "  <%= command.name.ljust maxlen %>   <%= command.summary %>"
