+++ lib/bashly/views/command/usage_fixed_flags.erb
@@ -1,7 +1,7 @@
 # :command.usage_fixed_flags
 echo "  --help, -h"
-echo -e "    <%= Settings.strings["show_this_help"] %>"
+echo -e "    <%= strings["show_this_help"] %>"
 echo
 echo "  --version"
-echo -e "    <%= Settings.strings["show_version_number"] %>"
+echo -e "    <%= strings["show_version_number"] %>"
 echo
